♻️ refactor: update import paths for dashboard components and adjust navbar links
